% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aminoacid_sequence.R
\name{aminoacid_sequence}
\alias{aminoacid_sequence}
\title{Title aminoacid_sequence}
\usage{
aminoacid_sequence(codons)
}
\arguments{
\item{codons}{sequence of codons that want to be converted into amino acids}
}
\value{
amino acid sequence
}
\description{
Title aminoacid_sequence
}
\examples{
# example code

codons_example <-  c("AUG", "UUU", "UUA", "UGG", "UCU")
aminoacid_sequence(codons_example)

}
