% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_five.R
\name{BarplotAAFreq}
\alias{BarplotAAFreq}
\title{BarplotAAFreq}
\usage{
BarplotAAFreq(AAS)
}
\arguments{
\item{AAS}{String of Aminoacids}
}
\value{
Variable that contains a barplot with the frequence of each of the aminoacids in the given string
}
\description{
The function takes a string of Aminoacids, splits it by characters and count how many times each character (aa) is present, then it plot the aas distribution in a barplot using ggplot
}
