% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon}
\alias{codon}
\title{Codon conversion from Codon to Amino Acid}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon
}
\description{
A dictionary whose keys are Codons and values are Amino Acids
}
\keyword{datasets}
