% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_plot.R
\name{col_plot}
\alias{col_plot}
\title{Column plot of the first amino acid in a codon}
\usage{
col_plot(peptide)
}
\arguments{
\item{peptide}{A variable which input a peptide#'}
}
\value{
A col plot describing the occurence of the first amino acid
}
\description{
Column plot of the first amino acid in a codon
}
