% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codons.R
\name{get_codons}
\alias{get_codons}
\title{Codon finder}
\usage{
get_codons(seq, start = 1)
}
\arguments{
\item{seq}{input sequence}

\item{start}{starting position of sequence reading frame}
}
\value{
A array of codons
}
\description{
Codon finder
}
\examples{
get_codons("ACTGAC")
}
