% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate}
\alias{replicate}
\title{generate random DNA sequence of a specified length}
\usage{
replicate(dna_seq_length)
}
\arguments{
\item{dna_seq_length}{length of random DNA sequence to be generated}
}
\value{
generated random DNA sequence
}
\description{
generate random DNA sequence of a specified length
}
\examples{
replicate(9)
}
