% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcribe}
\alias{transcribe}
\title{Transcription}
\usage{
transcribe(replication)
}
\arguments{
\item{replication}{replicated DNA sequence}
}
\value{
transcribed DNA sequence
}
\description{
Transcription
}
