% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate codons into amino acid sequence}
\usage{
translate(codons)
}
\arguments{
\item{codons}{a set of codons}
}
\value{
a sequence of amino acids
}
\description{
Translate codons into amino acid sequence
}
