% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_plot.R
\name{col_plot}
\alias{col_plot}
\title{Title}
\usage{
col_plot(codon_list)
}
\arguments{
\item{codon}{A variable which input a codon sequence}
}
\value{
A col plot describing the repartition of codons in the table
}
\description{
Title
}
\examples{
col_plot(codon)
}
